fixed init issues - you can now MOVE

+++ src/server/logic/messagehandling.h
@@ -13,9 +13,9 @@ int remove_client(Client clients[MAX_CLIENTS], int id);
 
 int srv_process_hello(GameState* game, ClientMessage* msg);
 int srv_process_disconnect(GameState* game, int fd);
-uint8_t srv_process_move_attempt(GameState* game, Message* msg);
-uint8_t srv_process_bomb_attempt(GameState* game, Message* msg);
-uint8_t srv_process_ready(GameState* game, Message* msg);
-uint8_t srv_start_game(GameState* game);
+int srv_process_move_attempt(GameState* game, Message* msg);
+int srv_process_bomb_attempt(GameState* game, Message* msg);
+int srv_process_ready(GameState* game, Message* msg);
+int srv_start_game(GameState* game);
 
 #endif
